Fixed compile time errors in hcf, refactores some packages
