+++ src/main/java/me/borawski/hcf/gui/PlayerPunishmentsGUI.java
@@ -1,4 +1,4 @@
-package me.borawski.hcf.gui.custom.edit;
+package me.borawski.hcf.gui;
 
 import org.bukkit.ChatColor;
 import org.bukkit.DyeColor;
@@ -6,10 +6,6 @@ import org.bukkit.Material;
 import org.bukkit.entity.Player;
 
 import me.borawski.hcf.Core;
-import me.borawski.hcf.gui.CustomIS;
-import me.borawski.hcf.gui.ItemGUI;
-import me.borawski.hcf.gui.MenuItem;
-import me.borawski.hcf.gui.custom.PlayerInfoGUI;
 import me.borawski.hcf.punishment.Punishment;
 import me.borawski.hcf.session.Session;
